requirements.txt
streamlit==1.36.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
matplotlib==3.9.2
scikit-image==0.24.0
scipy==1.14.1
Pillow==10.4.0